Initial support for DDR4 on Agilex 5 (#1654)
Add a Platform Designer project to wrap DDR4 controllers since Agilex 5 does
not have the memory subsystem used on Agilex 7. The Agilex 5 Dev Kit has two
DIMMs, one with ECC and one without. The project wraps both with the same
AXI-MM interface, ignoring the ECC bits.

The PD project exports interface with names and fields as close as possible
to the memory subsystem equivalents. The OFS SystemVerilog PD wrapper script
maps the exported ports to SV interfaces and vectors.

There are still known data corruption and throughput problems in the
implementation, to be fixed in later commits.